:recycle: split hold and recruit infomation panel
