+++ frontend/src/views/components/event/add/AddEventForm.tsx
@@ -4,10 +4,7 @@ import { FormValues } from '../../../forms/addEvent';
 import { FormikProps } from 'formik';
 import Button from '@material-ui/core/Button';
 import Select from 'react-select';
-import Input from '@material-ui/core/Input';
 import Grid from '@material-ui/core/Grid';
-import InputAdornment from '@material-ui/core/InputAdornment';
-import RangeDatePicker from '../../form/RangeDatePicker';
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
 import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
@@ -15,6 +12,7 @@ import Typography from '@material-ui/core/Typography';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import BasicInfomationPanel from './BasicInfomationsPanel';
 import ThrowingMoneyPanel from './ThrowingMoneyPanel';
+import HoldAndRecruitInfomationsPanel from './HoldAndRecruitInfomationsPanel';
 
 // グループを一括で取得する処理はまた別pullreqで対応予定
 const groups = [
@@ -28,13 +26,6 @@ const categories = [
   { value: 1, label: 'category1' },
   { value: 2, label: 'category2' },
   { value: 3, label: 'category3' }
-];
-
-// 開催者を一括で取得する処理はまた別pullreqで対応予定
-const organizers = [
-  { value: 1, label: 'organizer1' },
-  { value: 2, label: 'organizer2' },
-  { value: 3, label: 'organizer3' }
 ];
 
 interface OtherProps {
@@ -56,63 +47,7 @@ const AddEventForm: React.FC<OtherProps & FormikProps<FormValues>> = (
     <form onSubmit={handleSubmit}>
       <BasicInfomationPanel {...props} />
       <ThrowingMoneyPanel {...props} />
-      <ExpansionPanel defaultExpanded>
-        <ExpansionPanelSummary
-          expandIcon={<ExpandMoreIcon />}
-          aria-controls="panel1a-content"
-          id="panel1a-header"
-        >
-          <Typography>開催・募集情報</Typography>
-        </ExpansionPanelSummary>
-        <ExpansionPanelDetails>
-          <Grid container spacing={1}>
-            <Grid item xs={12}>
-              <Select
-                closeMenuOnSelect={false}
-                isMulti
-                value={values.organizers}
-                onChange={handleChange('organizers')}
-                options={organizers}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <Input
-                id="adornment-capacity"
-                value={values.capacity}
-                onChange={handleChange('capacity')}
-                endAdornment={
-                  <InputAdornment position="end">人</InputAdornment>
-                }
-                aria-describedby="capacity-helper-text"
-                inputProps={{
-                  'aria-label': 'capacity'
-                }}
-                fullWidth
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <RangeDatePicker
-                idPrefix="recruit"
-                labelPrefix="募集"
-                startDate={values.recruitStartDate}
-                endDate={values.recruitEndDate}
-                startDateOnChange={handleChange('recruitStartDate')}
-                endDateOnChange={handleChange('recruitEndDate')}
-              />
-            </Grid>
-            <Grid item xs={12}>
-              <RangeDatePicker
-                idPrefix="hold"
-                labelPrefix="開催"
-                startDate={values.holdStartDate}
-                endDate={values.holdEndDate}
-                startDateOnChange={handleChange('holdStartDate')}
-                endDateOnChange={handleChange('holdEndDate')}
-              />
-            </Grid>
-          </Grid>
-        </ExpansionPanelDetails>
-      </ExpansionPanel>
+      <HoldAndRecruitInfomationsPanel {...props} />
       <ExpansionPanel defaultExpanded>
         <ExpansionPanelSummary
           expandIcon={<ExpandMoreIcon />}
